update packages, move schema validation to apis, use new schema methods to create api
