+++ asset/asset.json
@@ -1,5 +1,5 @@
 {
     "name": "kafka",
     "description": "Kafka reader and writer support.",
-    "version": "3.0.1"
+    "version": "3.1.0"
 }
