+++ asset/package.json
@@ -1,6 +1,6 @@
 {
     "name": "kafka-assets",
-    "version": "3.0.1",
+    "version": "3.1.0",
     "private": true,
     "workspaces": {
         "nohoist": [
