+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "kafka-asset-bundle",
-    "version": "3.0.1",
+    "version": "3.1.0",
     "description": "A bundle of Kafka operations and processors for Teraslice",
     "private": true,
     "workspaces": {
@@ -27,11 +27,11 @@
     "devDependencies": {
         "@terascope/eslint-config": "^0.5.0",
         "@types/jest": "^26.0.14",
-        "@types/node": "^14.11.5",
+        "@types/node": "^14.11.8",
         "@types/uuid": "^8.3.0",
         "bunyan": "^1.8.14",
-        "eslint": "^7.10.0",
-        "jest": "^26.5.2",
+        "eslint": "^7.11.0",
+        "jest": "^26.5.3",
         "jest-extended": "^0.11.5",
         "teraslice-test-harness": "^0.26.0",
         "ts-jest": "^26.4.1",
