+++ test/kafka_sender/schema-spec.ts
@@ -1,6 +1,5 @@
 import 'jest-extended';
 import {
-    TestContext,
     newTestJobConfig,
     OpConfig,
     APIConfig,
@@ -9,13 +8,10 @@ import {
     Logger
 } from '@terascope/job-components';
 import { WorkerTestHarness } from 'teraslice-test-harness';
-import Schema from '../../asset/src/kafka_sender/schema';
 import { kafkaBrokers } from '../helpers/config';
 import Connector from '../../packages/terafoundation_kafka_connector/dist';
 
 describe('Kafka Sender Schema', () => {
-    const context = new TestContext('kafka-sender');
-    const schema = new Schema(context);
     const mockFlush = jest.fn();
 
     let harness: WorkerTestHarness;
@@ -78,34 +74,17 @@ describe('Kafka Sender Schema', () => {
         });
 
         it('should not throw an error if api and sender make a valid config', async () => {
-            const opConfig = { _op: 'kafka_sender' };
+            const opConfig = { _op: 'kafka_sender', api_name: 'kafka_sender_api', };
             const apiConfig = { _name: 'kafka_sender_api', topic: 'hello' };
 
             await expect(makeTest(opConfig, apiConfig)).toResolve();
         });
 
-        it('should throw an error if opConfig topic is specified and api is set', async () => {
+        it('should throw an error if opConfig topic is specified and api is set differently', async () => {
             const opConfig = { _op: 'kafka_sender', topic: 'hello', api_name: 'kafka_sender_api' };
-            const apiConfig = { _name: 'kafka_sender_api', topic: 'hello' };
+            const apiConfig = { _name: 'kafka_sender_api', topic: 'stuff' };
 
             await expect(makeTest(opConfig, apiConfig)).toReject();
-        });
-
-        it('should throw an error topic is specified and api is set when api is indirectly referenced', async () => {
-            const opConfig = { _op: 'kafka_sender', topic: 'hello' };
-            const apiConfig = { _name: 'kafka_sender_api', topic: 'hello' };
-
-            await expect(makeTest(opConfig, apiConfig)).toReject();
-        });
-
-        it('should set the required_acks default to 1', () => {
-            expect(schema.validate({
-                _op: 'kafka_sender',
-                topic: 'hello',
-                size: 1
-            })).toMatchObject({
-                required_acks: 1
-            });
         });
     });
 });
